% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxiety.R
\docType{data}
\name{anxiety}
\alias{anxiety}
\title{data set looking at treatment outcomes for anxiety}
\format{
\subsection{\code{anxiety}}{

A data frame with 100 rows and 4 columns:
\describe{
\item{Subject}{Subject ID}
\item{GrpNum}{1-5}
\item{Group}{Name of treatment group; Control, CBT+Drug, CBT+Placebo, Drug, Placebo}
\item{Improvement}{Amount of improvement}
...
}
}
}
\source{
Dr. Melanie Gleitzman
}
\usage{
anxiety
}
\description{
A simulated data set containing 100 observations
}
\keyword{datasets}
