% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_ph_b.R
\name{cc_between}
\alias{cc_between}
\title{Critical constant for post-hoc simultaneous between CIs}
\usage{
cc_between(v_b, v_e, alpha = 0.05)
}
\arguments{
\item{v_b}{\itemize{
\item a single numeric value - df between
}}

\item{v_e}{\itemize{
\item a single numeric value - df residual
}}

\item{alpha}{\itemize{
\item a single numeric value - acceptable type 1 error rate for
the family of contrasts
}}
}
\value{
\itemize{
\item a single numeric value that is the critical constant for use in the
construction of post-hoc simultaneous confidence intervals for between
subject effects
}
}
\description{
Computes the critical constant for use in calculating post-hoc simultaneous
confidence intervals for between groups contrast effects. For further
details see equation 7 (p. 217) from Bird (2002),
https://doi.org/10.1177/0013164402062002001.
}
\examples{
N = 16 # total number of subjects
J = 4 # number of groups
v_b = J-1
v_e = N - J
alpha = 0.05
cc_between(v_b, v_e, alpha)
}
