% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_ph_bw.R
\name{cc_ph_bw}
\alias{cc_ph_bw}
\title{Critical constant for post-hoc simultaneous between x within CIs}
\usage{
cc_ph_bw(v_w, v_b, v_e, alpha)
}
\arguments{
\item{v_w}{a single numeric value - df within}

\item{v_b}{a single numeric value - df between}

\item{v_e}{a single numeric value - df error}

\item{alpha}{desired type 1 error rate e.g. p=.05}
}
\value{
a single numeric value that is the critical constant for use in the
construction of post-hoc simultaneous confidence intervals for between x
within subject effects
}
\description{
Computes the critical constant for use in calculating post-hoc simultaneous
confidence intervals for between x within contrast effects. For further
details see equation 9 (p. 217) from Bird (2002),
https://doi.org/10.1177/0013164402062002001.
}
\examples{
N = 16 # total number of subjects
W = 3 # number of levels within
J = 4 # number of groups
v_w = W - 1
v_b = J - 1
v_e = N - J
alpha = 0.05
cc_ph_bw(v_w, v_b, v_e, alpha)
}
