% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_smr.R
\name{cc_smr}
\alias{cc_smr}
\title{Convert the critical constant for the SMR test into a critical constant for the
computation of confidence intervals. This takes the square root of the SMR parameter
output by smr_crit, as per Bird & Hadzi-Pavlovic (2005), page 360.}
\usage{
cc_smr(p, q, v_e, alpha, n_sim = 1e+05, seed = NULL)
}
\arguments{
\item{p}{Integer. The matrix dimension (number of variables).}

\item{q}{Integer. The degrees of freedom for the Wishart distribution.}

\item{alpha}{Numeric. The significance level (e.g., 0.05 for a 95\%
confidence level).}

\item{n_sim}{Integer. The number of simulations to run for estimating
the moments. Default is 100000.}

\item{seed}{Integer or NULL. The random seed for reproducibility.
Default is NULL.}

\item{n}{Integer. The degrees of freedom for the test statistic.}
}
\value{
Numeric. The critical value for the confidence interval computation.
}
\description{
Convert the critical constant for the SMR test into a critical constant for the
computation of confidence intervals. This takes the square root of the SMR parameter
output by smr_crit, as per Bird & Hadzi-Pavlovic (2005), page 360.
}
\examples{
critical_constant <- cc_smr(2, 2, 96, .05)
print(critical_constant)
}
