% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_contrast_ci.R
\name{compute_contrast_ci}
\alias{compute_contrast_ci}
\title{Compute the confidence interval for a contrast}
\usage{
compute_contrast_ci(cc, se_contrast)
}
\arguments{
\item{cc}{The critical constant, given the effect (see Bird(2002), eq 3, term CC)}

\item{se_contrast}{The standard error of the contrast (Bird(2002), eq 3, sigma)}
}
\value{
the range of confidence, to be added to/subtracted from the effect
}
\description{
Compute the confidence interval for a contrast
}
\examples{
cc <- 2.5
se_contrast <- 1.2
compute_contrast_ci(cc, se_contrast)
}
