% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_df.R
\name{compute_df}
\alias{compute_df}
\title{Compute degrees of freedom for the within or between family of contrasts}
\usage{
compute_df(model, fctrs)
}
\arguments{
\item{model}{an anova model object of the afex_aov class}

\item{fctrs}{a list of characters or strings - list of relevant factor names}
}
\value{
a single value which is the degrees of freedom for that family
}
\description{
Given a list of within or between factors, and an object of the afex_aov
model class, this function will compute the degrees of freedom for that
family For further details see p. 216-217 from Bird (2002),
https://doi.org/10.1177/0013164402062002001.
}
\examples{
data(spacing)
mod <- aov_ez("subj", "yield", spacing,
               within = "spacing", between = "group")
compute_df(mod, list("group")) # get the degrees of freedom for the
between family
compute_df(mod, list("spacing")) # get df for the within family

or aritrary example to show you can add more than one factor to the list:
mod <- aov_ez(some data and inputs)
fctrs <- list("groupA", "groupB", "groupC")
compute_df(mod, fctrs)
}
