% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression}
\alias{depression}
\title{depression - patient happiness scores from pre to post and to follow up}
\format{
\subsection{\code{depression}}{

A data frame with 45 rows and 5 columns:
\describe{
\item{Group}{Treatment Group: Tmt1 (treatment 1), Tmt2, Ctrl}
\item{GrpNum}{Numerical vrsn of Group (1, 2, or 3)}
\item{Time}{Timepoint of happiness measurement - Pre, Post, FU}
\item{Happiness}{Happiness score: higher is better}
\item{Subject}{Subject number}
...
}
}
}
\source{
<UNSW: Melanie Gleitzman>
}
\usage{
depression
}
\description{
Data showing happiness scores from 15 patients. Scores taken at Pre, Post and
Follow Up (FU) to a treatment programme. Patients are allocated to one of three
Treatment Groups (Tmt1, Tmt2, Ctrl).
}
\keyword{datasets}
