% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distraction.R
\docType{data}
\name{distraction}
\alias{distraction}
\title{data set looking at performance under different distraction conditions}
\format{
\subsection{\code{distraction}}{

A data frame of simulated data containing 90 rows and 4 columns
\describe{
\item{Subject}{Subject ID}
\item{Auditory}{Level of Auditory distraction; Low, Medium, or High}
\item{Visual}{Level of Visual distraction; None, Dissimilar, Similar
\item{Performance}{Performance Score}
...
}
}
}
\source{
Dr. Melanie Gleitzman
}
\usage{
distraction
}
\description{
data set looking at performance under different distraction conditions
}
\keyword{datasets}
