% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experience}
\alias{experience}
\title{experience - patient experience data}
\format{
\subsection{\code{experience}}{

A data frame with 132 rows and 5 columns:
\describe{
\item{Experience}{Yrs experience of therapist: 10, 3, or <1}
\item{Therapy}{Therapy type: Behavioural, Cognitive, Psychotherapy or Group}
\item{GrpNum}{Group number: 1-12}
\item{Outcome}{Patient wellbeing score: higher is better}
\item{SubjID}{Subject ID}
...
}
}
}
\source{
<UNSW: Dr. Melanie Gleitzman>
}
\usage{
experience
}
\description{
A study examining the effect of experience of therapist and type of therapy
on patient outcome. 132 patients are randomly allocated to one of 12 groups
in a 3 x 4 two-way factorial design (between-groups).
}
\keyword{datasets}
