% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatigue.R
\docType{data}
\name{fatigue}
\alias{fatigue}
\title{Fatigue data}
\format{
\subsection{\code{fatigue}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{Group}{Fatigue Group: 1=alert, 2=mild, 3=moderate, 4=extreme}
\item{Errors}{Number of driving errors}
\item{id}{Subject ID}
...
}
}
}
\source{
Dr. Melanie Gleitzman, UNSW
}
\usage{
fatigue
}
\description{
A simulated data set containing data on driver errors in four groups
experiencing differing levels of fatigue.
}
\keyword{datasets}
