% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcr_cdf.R
\name{gcr_cdf}
\alias{gcr_cdf}
\title{Cumulative density function for Roy's (1953) Greatest Characteristic Root
distribution}
\usage{
gcr_cdf(s, m, n, theta)
}
\arguments{
\item{s}{min(v_b, v_w) where v_b = df between, and v_w = df within (Bird,
2002, p. 217)}

\item{m}{(|v_b - v_w|-1)/2}

\item{n}{(v_e - v_w - 1)/2 where v_e = N-J, where N is the total sample
size and J is the number of groups}

\item{theta}{the critical GCR value, for which the probability of is desired}
}
\value{
a single numeric value that is p(theta)
}
\description{
Compute the cumulative density function for the largest root of a covariance
matrix, given s, m, n, and theta.
}
\details{
See Chiani et al (2016) for in relation to GCR:
http://dx.doi.org/10.1016/j.jmva.2015.10.007

and See Bird (2002) for definitions of s, m, & n
https://doi.org/10.1177/0013164402062002001
}
\examples{
# a 2 x 2 factorial design where factor A is between and B is within, w
# 100 participants
s <- 1
m <- -0.5
n < 96
theta <- 0.4
gcr_cdf(s,m,n,theta)

}
