% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcr_crit.R
\name{gcr_crit}
\alias{gcr_crit}
\title{Find the critical value theta (GCR), given alpha, s, m, & n}
\usage{
gcr_crit(alpha, s, m, n)
}
\arguments{
\item{alpha}{alpha/type 1 error rate - e.g. .05}

\item{s}{min(v_b, v_w) where v_b = df between, and v_w = df within (Bird,
2002, p. 217)}

\item{m}{(|v_b - v_w|-1)/2}

\item{n}{(v_e - v_w - 1)/2 where v_e = N-J, where N is the total sample
size and J is the number of groups}
}
\value{
a single value, giving the critical theta/GCR expected under the
null hypothesis (given the covariance matrix)
}
\description{
This function takes a given alpha rate (e.g. p=.05), and finds the
corresponding critical value of theta - i.e. the greatest characteristic root
that would be expected p=0.05 of the time, if the null hypothesis is true,
and given the parameter values of s,m, and n (see below).
The greatest characteristic root (GCR) is the largest linear combination of
effects from a covariance matrix, and can be used as the critical value
used to compute confidence intervals for between x within ANOVA designs.
}
\details{
For further info see:
Bird (2002) https://doi.org/10.1177/0013164402062002001
Bird & Hadzi-Pavlovic (1983) https://doi.org/10.1037/0033-2909.93.1.167
Pillai (1965) https://www.jstor.org/stable/2333693
}
\examples{
gcr_crit(.05, 1, -0.5, 96)

}
