% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cis.R
\name{get_cis}
\alias{get_cis}
\title{Calculate CIs for a contrast table (emmeans object) given a critical constant}
\usage{
get_cis(contrast_table, cc)
}
\arguments{
\item{contrast_table}{\itemize{
\item a table of contrasts, produced using the 'contrast'
function from emmeans
}}

\item{cc}{\itemize{
\item a critical constrant, to be used to calculate CIs.
}}
}
\value{
contrast_table - a summary table of the contrasts with the CI info added
}
\description{
Calculate CIs for a contrast table (emmeans object) given a critical constant
}
\examples{
data(spacing)
# fit model
mod <- afex::aov_ez("subj","yield", spacing, within = "spacing", between = "group")
sum_emm_btwn <- emmeans::emmeans(mod, "group") # get marginal means
con <- list( # define some contrasts
            "12vs34" = c(0.5, 0.5, -0.5, -0.5),
             "1vs2" = c(1, -1, 0, 0),
             "3vs4" = c(0, 0, 1, -1)
             )
btwn_con <- emmeans::contrast(sum_emm_btwn, con) # get contrast table
cc <- cc_ind_t(v_e=12, alpha=0.05) # for example
get_cis(btwn_con, cc)
}
