% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrast_coefficients.R
\name{get_contrast_coefficients}
\alias{get_contrast_coefficients}
\title{Get the contrast coefficient vectors from a list of contrast tables output by
emmeans \code{contrast} function}
\usage{
get_contrast_coefficients(contrast_tables)
}
\arguments{
\item{contrast_tables}{}
}
\value{
a list of the same size as contrast tables. Each element of the list
is a c x k matrix of contrast coefficient vectors
}
\description{
Get the contrast coefficient vectors from a list of contrast tables output by
emmeans \code{contrast} function
}
\examples{
data(spacing)
# fit model
mod <- afex::aov_ez("subj","yield", spacing, within = "spacing", between = "group")
sum_emm_btwn <- emmeans::emmeans(mod, "group") # get marginal means
con <- list( # define some contrasts
            "12vs34" = c(0.5, 0.5, -0.5, -0.5),
             "1vs2" = c(1, -1, 0, 0),
             "3vs4" = c(0, 0, 1, -1)
             )
btwn_con <- emmeans::contrast(sum_emm_btwn, con) # get contrast table
get_contrast_coefficients(btwn_con)
}
