% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels.R
\name{get_levels}
\alias{get_levels}
\title{get number of levels in a factor}
\usage{
get_levels(data, fctr)
}
\arguments{
\item{data}{a longform dataframe. Predictor variables should be factors}

\item{factor}{a single string which matches the name of the factor in the
dataframe}
}
\value{
a single numeric value which is the number of levels in factor fctr
}
\description{
from a long form data frame, get the number of levels
in the factor specified by factor
}
\examples{
data = data.frame(A = c("X", "X", "X", "Y", "Y", "Y"))
data$A <- as.factor(data$A)
fctr = "A"
get_levels(data, fctr)

}
