% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kevin_within}
\alias{kevin_within}
\title{kevin_within - small set of scores from pre to post and to follow up}
\format{
\subsection{\code{kevin_within}}{

A data frame with 5 rows and 4 columns:
\describe{
\item{Subject}{Subject Number: 1-5}
\item{Pre}{Score at Pre}
\item{Post}{Score at Post}
\item{FU}{Score at follow up (FU)}
...
}
}
}
\source{
<Bird (2004) Analysis of Variance via Confidence Intervals>
}
\usage{
kevin_within
}
\description{
Data showing scores from 5 participants. Scores taken at Pre, Post and
Follow Up (FU). Dataset appears on page 123 of Bird (2004)
https://methods.sagepub.com/book/mono/analysis-of-variance-via-confidence-intervals/toc
}
\keyword{datasets}
