% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priming}
\alias{priming}
\title{Priming lexical decision task data}
\format{
\subsection{\code{priming}}{

A data frame with 15 rows and 4 columns:
\describe{
\item{subj}{subject ID}
\item{Condition}{within-subjects factor with 3 levels: STEREO, ATYPICAL, NEUTRAL}
\item{WordType}{same as Condition, possibly labeled differently}
\item{RT}{reaction time in tenths of a second}
}
}
}
\source{
Dr. Melanie Gleitzman, UNSW
}
\usage{
priming
}
\description{
A synthetic dataset modeling a lexical decision task
used to investigate automatic activation of stereotypes through priming.
Participants are subliminally primed with a social category label
(e.g., ‘unemployed’ or ‘elderly’) before responding to target words.
The target words are either stereotypical attributes, atypical attributes,
or neutral words. The dependent variable is reaction time (in tenths of a second).
}
\details{
Priming is hypothesized to enhance recognition for stereotypical attributes
(i.e., faster reaction times), and inhibit recognition for atypical attributes,
relative to neutral words.
}
\keyword{datasets}
