% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_cont.R
\name{se_cont}
\alias{se_cont}
\title{Compute the standard error of a contrast}
\usage{
se_cont(MSE, contrast_vector, n)
}
\arguments{
\item{MSE}{mean square error: a single numeric value}

\item{contrast_vector}{a vector defining the contrast}

\item{n}{n observations: a single numeric value}
}
\value{
SE: a single numeric value
}
\description{
This computes the estimated standard error of
a contrast, given the mean square error (MSE),
the contrast vector, and the relevant n.
}
\details{
See the definition of sigma from equation 3 of
Bird (2002), p. 203
}
\examples{
x <- 2.2
y <- c(1, -1)
z <- 50
se_cont(x, y, z)
}
