% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_crit.R
\name{smr_crit}
\alias{smr_crit}
\title{Compute the critical value for the Studentized Maxmimum Root (SMR)
distribution}
\usage{
smr_crit(alpha, p, q, n, n_sim = 1e+05, seed = NULL)
}
\arguments{
\item{alpha}{Numeric. The significance level (e.g., 0.05 for a 95\%
confidence level).}

\item{p}{Integer. The matrix dimension (number of variables).}

\item{q}{Integer. The degrees of freedom for the Wishart distribution.}

\item{n}{Integer. The degrees of freedom for the test statistic.}

\item{n_sim}{Integer. The number of simulations to run for estimating
the moments. Default is 100000.}

\item{seed}{Integer or NULL. The random seed for reproducibility.
Default is NULL.}
}
\value{
Numeric. The critical value for the statistical test.
}
\description{
The \code{smr_crit} function calculates the critical value for a
statistical test using the moments of the largest roots of Wishart
matrices and an approximation to the F-distribution.
}
\details{
The function first simulates the largest roots of Wishart matrices
using the \code{wishartlr_sample} function to estimate the first three
moments (mean, variance, and skewness). These moments are then used
to approximate the parameters of an F-distribution, which is used to
compute the critical value for the given significance level
(\code{alpha}).
}
\examples{
# Compute critical value for a test
crit_value <- smr_crit(alpha = 0.05, p = 5, q = 10, n = 20, n_sim = 100000, seed = 123)
print(crit_value)

}
\seealso{
wishartlr_sample
}
