% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spacing}
\alias{spacing}
\title{Plant spacing data}
\format{
\subsection{\code{spacing}}{

A data frame with 48 rows and 4 columns:
\describe{
\item{subj}{subject}
\item{group}{between subjects factor - 1:4}
\item{spacing}{within subjects factor - TWENTY, FORTY, or SIXTY}
\item{yield}{number of plants yielded}
...
}
}
}
\source{
\url{https://www.unsw.edu.au/science/our-schools/psychology/our-research/research-tools/psy-statistical-program}
}
\usage{
spacing
}
\description{
A small dataset for testing PsyR functionality. Data has plant yield
as the DV, and comes from a 4 (group) x 3 (spacing) mixed design.
}
\keyword{datasets}
