% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishartlr_sample.R
\name{wishartlr_sample}
\alias{wishartlr_sample}
\title{Simulate largest roots of Wishart matrices}
\usage{
wishartlr_sample(p, q, n_sim = 1e+05, return_samples = FALSE, seed = NULL)
}
\arguments{
\item{p}{Integer. The first matrix dimension.}

\item{q}{Integer. The second matrix dimension.}

\item{n_sim}{Integer. The number of simulations to run. Default is
10000.}

\item{seed}{Integer or NULL. The random seed for reproducibility.
Default is NULL.}
}
\value{
A list containing the following elements:
\item{mu1}{The mean of the largest roots.}
\item{mu2}{The second moment (variance) of the largest roots. }
\item{mu3}{The third moment (skewness) of the largest roots.}
\item{largest_roots}{A numeric vector of the largest roots simulated.
Only returned if \code{return_samples} is TRUE.}
}
\description{
The \code{wishartlr_sample} function simulates the largest roots of
Wishart matrices given the matrix dimensions and number of
simulations and calculates the first three moments of the resulting
distribution.
}
\details{
The function generates Wishart matrices with an identity
scale matrix and computes the largest eigenvalue (root) of each
matrix. The Wishart distribution is commonly used in multivariate
statistics, particularly in covariance matrix estimation.
}
\examples{
# Simulate largest roots of Wishart matrices
wishart_moments <- wishartlr_sample(p = 5, q = 10, n_sim = 1000, seed =
123)

}
