% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psyci.R
\name{psyci}
\alias{psyci}
\title{Calculate confidence intervals for contrasts of interest from ANOVA models}
\usage{
psyci(
  model,
  contrast_tables,
  method,
  between_factors = NA,
  within_factors = NA,
  alpha = 0.05
)
}
\arguments{
\item{model}{an anova model made using the afex package (class afex_aov)}

\item{contrast_tables}{a LIST of contrast tables generated using emmeans contrast()}

\item{method}{a single string - method for confidence interval computation -
"ind", "bf", or "ph" - see further details below}

\item{between_factors}{\itemize{
\item a list of between subject factor names, default = NA
}}

\item{within_factors}{\itemize{
\item a list of within subject factor names, default = NA
}}

\item{alpha}{\itemize{
\item desired type 1 error rate
}}
}
\value{
an emmeans contrast table with confidence intervals added
}
\description{
This function computes confidence intervals for contrasts testing effects
from ANOVA models. The method used to compute the confidence intervals
is selected by the user (see below). The methods are described in Bird
(2002) (see below) https://doi.org/10.1177/0013164402062002001
}
\section{METHODS}{
ind - individual t ci's - A set of individual 100(1 - α)\% confidence
intervals on planned contrasts controls the noncoverage error rate at
α for each contrast in the set, thereby producing a familywise error
rate which is greater than α (see Bird, 2002, bottom of page 204).

bf - bonferonni t ci's - only valid if contrasts are defined independently
of the data. Controls the familywise non-coverage error rate at alpha, given
distributional assumptions are met. (see Bird, 2002, pg. 205)

ph - post-hoc ci's - valid for post-hoc contrasts, will control the
familywise non-coverage error rate at alpha, given distributional assumptions
are met (see Bird, 2002, pg. 217, eqs. 7-9).
}

\examples{
data(spacing)
# fit model
mod <- afex::aov_ez("subj","yield", spacing, within = "spacing", between = "group")
sum_emm_btwn <- emmeans::emmeans(mod, "group") # get marginal means
con <- list( # define some contrasts
            "12vs34" = c(0.5, 0.5, -0.5, -0.5),
             "1vs2" = c(1, -1, 0, 0),
             "3vs4" = c(0, 0, 1, -1)
             )
btwn_con <- emmeans::contrast(sum_emm_btwn, con) # get contrast table
# now add CI's using post-hoc method
psyci(model = mod, contrast_tables = list(btwn_con), method = "ph",
        family = "b", between_factors = list("group"))
}
